+++ test/lwm2mtestapplication/lwm2mtest.h
@@ -96,7 +96,7 @@ public:
                                                 uint16_t resource_instance = 0);
 
     bool create_static_resource_instance_int(const char *name,
-                                             int64_t value,
+                                             int32_t value,
                                              bool multiple_instance,
                                              uint16_t object_instance = 0,
                                              uint16_t resource_instance = 0);
